feat(reth-bench): add reporting and wait options to replay-payloads
Add same reporting functionality as new-payload-fcu:
- CombinedResult and TotalGasRow tracking per payload
- CSV output via --output flag
- Total Ggas/s summary at the end

Add wait-time and wait-for-persistence options:
- --wait-time for fixed sleep between blocks
- --wait-for-persistence for subscription-based waiting
- --persistence-threshold and --ws-rpc-url options

Extract shared persistence waiting code into persistence_waiter.rs:
- PersistenceWaiter, PersistenceSubscription structs
- engine_url_to_ws_url, setup_persistence_subscription functions
- PERSISTENCE_CHECKPOINT_TIMEOUT constant
